Patch to read XML with either "Plate" or "Group"

+++ artel_parse/artel_reader_v2.py
@@ -23,26 +23,36 @@ class XMLFormatter:
         self.window.destroy()
 
     def format_data(self):
-        for group in self.root.iter("Group"):
-            col_list = [col_num.text for column in group.iter("Columns") for col_num in column.iter('Name')]
-            col_len = len(col_list)
-            row_list = sorted(list(set([row.text for row in group.iter("Row")])))
-            row_len = len(row_list)
-            value_list = np.array([value.text for value in group.iter("Value")])
-            value_matrix = value_list.reshape(row_len, col_len)
-            group_df = pd.DataFrame(value_matrix, columns=col_list, index=row_list)
-            self.all_groups_df = pd.concat([self.all_groups_df, group_df], axis=1)
+        id_list = [tag.tag for tag in self.root]
+        if "Group" in id_list:
+            run_id = "Group"
+        elif "Plate" in id_list:
+            run_id = "Plate"
+        else:
+            run_id = ""
+        if run_id == "":
+            messagebox.showerror(title="Warning...", message="Something may have changed with the XML file.")
+        else:
+            for group in self.root.iter(run_id):
+                col_list = [col_num.text for column in group.iter("Columns") for col_num in column.iter('Name')]
+                col_len = len(col_list)
+                row_list = sorted(list(set([row.text for row in group.iter("Row")])))
+                row_len = len(row_list)
+                value_list = np.array([value.text for value in group.iter("Value")])
+                value_matrix = value_list.reshape(row_len, col_len)
+                group_df = pd.DataFrame(value_matrix, columns=col_list, index=row_list)
+                self.all_groups_df = pd.concat([self.all_groups_df, group_df], axis=1)
 
-        display(self.show_df, ipw.HTML("<h3><b><u>Output display</u></b></h3>"), self.all_groups_df)
-        file_open = True
-        while file_open:
-            try:
-                pd.DataFrame.to_excel(self.all_groups_df, f"{self.drop_path}/Artel_data_{self.now}.xlsx")
-            except PermissionError:
-                messagebox.showerror(title="Oops...", message="Do you have the file open?")
-            else:
-                messagebox.showinfo(title="Way to go!", message="Data has been formatted and output!")
-                file_open = False
+            display(self.show_df, ipw.HTML("<h3><b><u>Output display</u></b></h3>"), self.all_groups_df)
+            file_open = True
+            while file_open:
+                try:
+                    pd.DataFrame.to_excel(self.all_groups_df, f"{self.drop_path}/Artel_data_{self.now}.xlsx")
+                except PermissionError:
+                    messagebox.showerror(title="Oops...", message="Do you have the file open?")
+                else:
+                    messagebox.showinfo(title="Way to go!", message="Data has been formatted and output!")
+                    file_open = False
 
 
 if __name__ == "__main__":
